Add cryptid by id refactor
